pyproject.toml
[project]
name = "pdf-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mohammed Ali Bellary", email = "b.mohammedali@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.1.0",
    "aiofiles>=24.1.0",
    "boto3>=1.39.11",
    "botocore>=1.39.11",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "watchtower>=3.4.0",
]

[project.scripts]
worker = "pdf_processor.worker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/pdf_processor"]